+++ languages/elixir/exercises/concept/list-comprehensions/.meta/design.md
@@ -1,0 +1,24 @@
+## Learning objectives
+
+- Know what a _comprehension_ is
+- Know the three parts to a comprehension
+  - generator
+  - filter
+  - collectable
+- Using the comprehension to create a cartesian product
+
+## Out of scope
+
+- comprehension :into
+- bitstring generators
+
+## Concepts
+
+- `list-comprehensions`
+
+## Prerequisites
+
+- `lists`
+- `maps`
+- `enum`
+- `tuples`
